requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langfuse>=2.0.0
python-dotenv>=1.0.0
tenacity>=9.0.0
pydantic>=2.0.0
